Cleanup on CLI app, sync PHPCS
