+++ cli/CliApp/Command/Get/Labels.php
@@ -20,16 +20,12 @@ use JTracker\Container;
 class Labels extends Get
 {
 	/**
-	 * Constructor.
+	 * The command "description" used for help texts.
 	 *
-	 * @since   1.0
+	 * @var    string
+	 * @since  1.0
 	 */
-	public function __construct()
-	{
-		parent::__construct();
-
-		$this->description = 'Retrieve project labels from GitHub.';
-	}
+	protected $description = 'Retrieve project labels from GitHub.';
 
 	/**
 	 * Execute the command.
@@ -52,11 +48,11 @@ class Labels extends Get
 	}
 
 	/**
-	 * Get the projects labels.
+	 * Get the project labels.
 	 *
-	 * @return $this
+	 * @return  $this
 	 *
-	 * @since  1.0
+	 * @since   1.0
 	 */
 	protected function processLabels()
 	{
