+++ cli/CliApp/Command/Get/Project.php
@@ -16,16 +16,12 @@ namespace CliApp\Command\Get;
 class Project extends Get
 {
 	/**
-	 * Constructor.
+	 * The command "description" used for help texts.
 	 *
-	 * @since   1.0
+	 * @var    string
+	 * @since  1.0
 	 */
-	public function __construct()
-	{
-		parent::__construct();
-
-		$this->description = 'Get the whole project info from GitHub, including issues and issue comments.';
-	}
+	protected $description = 'Get the whole project info from GitHub, including issues and issue comments.';
 
 	/**
 	 * Execute the command.
@@ -66,9 +62,9 @@ class Project extends Get
 	/**
 	 * Process the project labels.
 	 *
-	 * @return $this
+	 * @return  $this
 	 *
-	 * @since  1.0
+	 * @since   1.0
 	 */
 	protected function processLabels()
 	{
@@ -81,9 +77,9 @@ class Project extends Get
 	/**
 	 * Process the project labels.
 	 *
-	 * @return $this
+	 * @return  $this
 	 *
-	 * @since  1.0
+	 * @since   1.0
 	 */
 	protected function processMilestones()
 	{
@@ -96,9 +92,9 @@ class Project extends Get
 	/**
 	 * Process the project issues.
 	 *
-	 * @return $this
+	 * @return  $this
 	 *
-	 * @since  1.0
+	 * @since   1.0
 	 */
 	protected function processIssues()
 	{
@@ -111,9 +107,9 @@ class Project extends Get
 	/**
 	 * Process the project comments.
 	 *
-	 * @return $this
+	 * @return  $this
 	 *
-	 * @since  1.0
+	 * @since   1.0
 	 */
 	protected function processComments()
 	{
@@ -126,9 +122,9 @@ class Project extends Get
 	/**
 	 * Process the project events.
 	 *
-	 * @return $this
+	 * @return  $this
 	 *
-	 * @since  1.0
+	 * @since   1.0
 	 */
 	protected function processEvents()
 	{
@@ -141,9 +137,9 @@ class Project extends Get
 	/**
 	 * Process the project avatars.
 	 *
-	 * @return $this
+	 * @return  $this
 	 *
-	 * @since  1.0
+	 * @since   1.0
 	 */
 	protected function processAvatars()
 	{
