+++ cli/CliApp/Command/Make/Dbcomments.php
@@ -11,23 +11,19 @@ namespace CliApp\Command\Make;
 use JTracker\Container;
 
 /**
- * Class for retrieving issues from GitHub for selected projects
+ * Class for generating class doc blocks in JTracker\Database\AbstractDatabaseTable classes
  *
  * @since  1.0
  */
 class Dbcomments extends Make
 {
 	/**
-	 * Constructor.
+	 * The command "description" used for help texts.
 	 *
-	 * @since   1.0
+	 * @var    string
+	 * @since  1.0
 	 */
-	public function __construct()
-	{
-		parent::__construct();
-
-		$this->description = 'Generate file headers for Table classes.';
-	}
+	protected $description = 'Generate file headers for Table classes';
 
 	/**
 	 * Execute the command.
@@ -98,7 +94,9 @@ class Dbcomments extends Make
 	 *
 	 * @param   array  $lines  The doc comment.
 	 *
-	 * @return \stdClass
+	 * @return  \stdClass
+	 *
+	 * @since   1.0
 	 */
 	private function getMaxVals(array $lines)
 	{
@@ -127,7 +125,9 @@ class Dbcomments extends Make
 	 *
 	 * @param   string  $type  The SQL data type.
 	 *
-	 * @return string
+	 * @return  string
+	 *
+	 * @since   1.0
 	 */
 	private function getType($type)
 	{
