+++ cli/CliApp/Command/Update/Update.php
@@ -2,8 +2,8 @@
 /**
  * Part of the Joomla! Tracker application.
  *
- * @copyright   Copyright (C) 2012 - 2013 Open Source Matters, Inc. All rights reserved.
- * @license     GNU General Public License version 2 or later; see LICENSE.txt
+ * @copyright  Copyright (C) 2012 - 2013 Open Source Matters, Inc. All rights reserved.
+ * @license    GNU General Public License version 2 or later; see LICENSE.txt
  */
 
 namespace CliApp\Command\Update;
@@ -27,36 +27,20 @@ use JTracker\Container;
 class Update extends TrackerCommand
 {
 	/**
+	 * The command "description" used for help texts.
+	 *
+	 * @var    string
+	 * @since  1.0
+	 */
+	protected $description = 'Used to update GitHub data';
+
+	/**
 	 * Joomla! Github object
 	 *
 	 * @var    Github
 	 * @since  1.0
 	 */
 	protected $github;
-
-	/**
-	 * Projects table object.
-	 *
-	 * @var    ProjectsTable
-	 * @since  1.0
-	 */
-	protected $project = null;
-
-	/**
-	 * Use the progress bar.
-	 *
-	 * @var    boolean
-	 * @since  1.0
-	 */
-	protected $usePBar;
-
-	/**
-	 * Progress bar format.
-	 *
-	 * @var    string
-	 * @since  1.0
-	 */
-	protected $pBarFormat = '[%bar%] %fraction% %elapsed% ETA: %estimate%';
 
 	/**
 	 * Constructor.
@@ -66,8 +50,6 @@ class Update extends TrackerCommand
 	public function __construct()
 	{
 		parent::__construct();
-
-		$this->description = 'Used to update GitHub data.';
 
 		$this
 			->addOption(
@@ -104,82 +86,6 @@ class Update extends TrackerCommand
 	}
 
 	/**
-	 * Select the project.
-	 *
-	 * @return  $this
-	 *
-	 * @since   1.0
-	 * @throws  \RuntimeException
-	 * @throws  AbortException
-	 * @todo    this might go to a base class.
-	 */
-	protected function selectProject()
-	{
-		$projects = with(new ProjectsModel(Container::getInstance()->get('db')))->getItems();
-
-		$id = $this->application->input->getInt('project', $this->application->input->getInt('p'));
-
-		if (!$id)
-		{
-			$this->out()
-				->out('<b>Available projects:</b>')
-				->out();
-
-			$cnt = 1;
-
-			$checks = array();
-
-			foreach ($projects as $project)
-			{
-				if ($project->gh_user && $project->gh_project)
-				{
-					$this->out('  <b>' . $cnt . '</b> (id: ' . $project->project_id . ') ' . $project->title);
-					$checks[$cnt] = $project;
-					$cnt++;
-				}
-			}
-
-			$this->out()
-				->out('<question>Select a project:</question> ', false);
-
-			$resp = (int) trim($this->application->in());
-
-			if (!$resp)
-			{
-				throw new AbortException('Aborted');
-			}
-
-			if (false == array_key_exists($resp, $checks))
-			{
-				throw new AbortException('Invalid project');
-			}
-
-			$this->project = $checks[$resp];
-		}
-		else
-		{
-			foreach ($projects as $project)
-			{
-				if ($project->project_id == $id)
-				{
-					$this->project = $project;
-
-					break;
-				}
-			}
-
-			if (is_null($this->project))
-			{
-				throw new AbortException('Invalid project');
-			}
-		}
-
-		$this->logOut('Processing project: <info>' . $this->project->title . '</info>');
-
-		return $this;
-	}
-
-	/**
 	 * Setup the Github object.
 	 *
 	 * @return  $this
@@ -192,33 +98,5 @@ class Update extends TrackerCommand
 		$this->github = Container::retrieve('gitHub');
 
 		return $this;
-	}
-
-	/**
-	 * Display the GitHub rate limit.
-	 *
-	 * @return  $this
-	 *
-	 * @since   1.0
-	 */
-	protected function displayGitHubRateLimit()
-	{
-		$this->application->displayGitHubRateLimit();
-
-		return $this;
-	}
-
-	/**
-	 * Get a progress bar object.
-	 *
-	 * @param   integer  $targetNum  The target number.
-	 *
-	 * @return  \Elkuku\Console\Helper\ConsoleProgressBar
-	 *
-	 * @since   1.0
-	 */
-	protected function getProgressBar($targetNum)
-	{
-		return $this->application->getProgressBar($targetNum);
 	}
 }
